Work in progress on doors
Right now it finds all potential places for a door (ie. shared wall segments between two rooms) and just makes them a door so that I could test it.  Ideally we want to have only one door between each pair of adjacent rooms, in such a way that the entire map is reachable.

Also there needs to be a way to animate the door opening/closing.