Modified images so they would repeat better
There are still some spots where you can see the seams, but I think it's
better now than before. Also looked at the tile, but that seems to work
well - just in the game you can see the repeat, so I'm not quite sure
why that is. Not a huge deal though as it looks pretty nice as is.